Modified ContainerFiles
