--- examples/kit-nextjs-article-starter/src/components/container/container-303030/container-303030.props.tsx
@@ -1,14 +1,0 @@
-import { PlaceholderProps } from 'types/Placeholder.props';
-import { ComponentProps } from '@/lib/component-props';
-
-import type { JSX } from 'react';
-
-/**
- * Model used for Sitecore Component integration
- */
-export type Container303030Props = ComponentProps &
-  PlaceholderProps & {
-    left?: JSX.Element;
-    center?: JSX.Element;
-    right?: JSX.Element;
-  };
