+++ examples/kit-nextjs-article-starter/src/components/container/container-3070/Container3070.tsx
@@ -1,13 +1,23 @@
 import type React from 'react';
 import { Placeholder, useSitecore } from '@sitecore-content-sdk/nextjs';
-import { Container3070Props } from '@/components/container/container-3070/container-3070.props';
 import {
   getContainerPlaceholderProps,
   isContainerPlaceholderEmpty,
 } from '@/components/container/container.util';
 import { Flex, FlexItem } from '@/components/flex/Flex.dev';
 import { cn } from '@/lib/utils';
+import { PlaceholderProps } from 'types/Placeholder.props';
+import { ComponentProps } from '@/lib/component-props';
+import type { JSX } from 'react';
 
+/**
+ * Model used for Sitecore Component integration
+ */
+type Container3070Props = ComponentProps &
+  PlaceholderProps & {
+    left?: JSX.Element;
+    right?: JSX.Element;
+  };
 export const Default: React.FC<Container3070Props> = (props) => {
   const { rendering, left, right } = props;
