--- examples/kit-nextjs-article-starter/src/components/container/container-4060/container-4060.props.tsx
@@ -1,13 +1,0 @@
-import { PlaceholderProps } from 'types/Placeholder.props';
-import { ComponentProps } from '@/lib/component-props';
-
-import type { JSX } from 'react';
-
-/**
- * Model used for Sitecore Component integration
- */
-export type Container4060Props = ComponentProps &
-  PlaceholderProps & {
-    left?: JSX.Element;
-    right?: JSX.Element;
-  };
