--- examples/kit-nextjs-article-starter/src/components/container/container-6040/container-6040.props.tsx
@@ -1,13 +1,0 @@
-import { PlaceholderProps } from 'types/Placeholder.props';
-import { ComponentProps } from '@/lib/component-props';
-
-import type { JSX } from 'react';
-
-/**
- * Model used for Sitecore Component integration
- */
-export type Container6040Props = ComponentProps &
-  PlaceholderProps & {
-    left?: JSX.Element;
-    right?: JSX.Element;
-  };
