+++ examples/kit-nextjs-article-starter/src/components/container/container-70/Container70.tsx
@@ -1,8 +1,20 @@
 import type React from 'react';
 import { Placeholder, useSitecore } from '@sitecore-content-sdk/nextjs';
-import { Container70Props } from '@/components/container/container-70/container-70.props';
 import { Flex, FlexItem } from '@/components/flex/Flex.dev';
 import { cn } from '@/lib/utils';
+import { PlaceholderProps } from '@/types/Placeholder.props';
+import { ComponentProps } from '@/lib/component-props';
+
+/**
+ * Model used for Sitecore Component integration
+ */
+type Container70Props = ComponentProps & PlaceholderProps & Container70Params;
+
+type Container70Params = {
+  params?: {
+    excludeTopMargin?: string;
+  };
+};
 
 export const Default: React.FC<Container70Props> = (props) => {
   const { rendering, children } = props;
