--- examples/kit-nextjs-article-starter/src/components/container/container-70/container-70.props.tsx
@@ -1,13 +1,0 @@
-import { PlaceholderProps } from '@/types/Placeholder.props';
-import { ComponentProps } from '@/lib/component-props';
-
-/**
- * Model used for Sitecore Component integration
- */
-export type Container70Props = ComponentProps & PlaceholderProps & Container70Params;
-
-type Container70Params = {
-  params?: {
-    excludeTopMargin?: string;
-  };
-};
