+++ examples/kit-nextjs-article-starter/src/components/container/container-7030/Container7030.tsx
@@ -1,12 +1,23 @@
 import type React from 'react';
 import { Placeholder, useSitecore } from '@sitecore-content-sdk/nextjs';
-import { Container7030Props } from '@/components/container/container-7030/container-7030.props';
 import {
   getContainerPlaceholderProps,
   isContainerPlaceholderEmpty,
 } from '@/components/container/container.util';
 import { Flex, FlexItem } from '@/components/flex/Flex.dev';
 import { cn } from '@/lib/utils';
+import { PlaceholderProps } from 'types/Placeholder.props';
+import { ComponentProps } from '@/lib/component-props';
+import type { JSX } from 'react';
+
+/**
+ * Model used for Sitecore Component integration
+ */
+type Container7030Props = ComponentProps &
+  PlaceholderProps & {
+    left?: JSX.Element;
+    right?: JSX.Element;
+  };
 
 export const Default: React.FC<Container7030Props> = (props) => {
   const { rendering, left, right } = props;
