+++ examples/kit-nextjs-article-starter/src/components/container/container-full-bleed/ContainerFullBleed.tsx
@@ -1,7 +1,24 @@
-import { ContainerFullBleedProps } from '@/components/container/container-full-bleed/container-full-bleed.props';
 import { Flex, FlexItem } from '@/components/flex/Flex.dev';
 import { Placeholder } from '@sitecore-content-sdk/nextjs';
 import { cva } from 'class-variance-authority';
+import { PlaceholderProps } from '@/types/Placeholder.props';
+import { ComponentProps } from '@/lib/component-props';
+import { BackgroundColor } from '@/enumerations/BackgroundColor.enum';
+
+/**
+ * Model used for Sitecore Component integration
+ */
+type ContainerFullBleedProps = ComponentProps & PlaceholderProps & ContainerFullBleedParams;
+
+type ContainerFullBleedParams = {
+  params?: {
+    backgroundColor?: BackgroundColor;
+    backgroundImagePath?: string;
+    excludeTopMargin?: string;
+    inset?: string;
+    [key: string]: any; // eslint-disable-line @typescript-eslint/no-explicit-any
+  };
+};
 
 export const Default: React.FC<ContainerFullBleedProps> = (props) => {
   const { rendering } = props;
