--- examples/kit-nextjs-article-starter/src/components/container/container-full-width/container-full-width.props.tsx
@@ -1,13 +1,0 @@
-import { PlaceholderProps } from '@/types/Placeholder.props';
-import { ComponentProps } from '@/lib/component-props';
-
-/**
- * Model used for Sitecore Component integration
- */
-export type ContainerFullWidthProps = ComponentProps & PlaceholderProps & ContainerFullWidthParams;
-
-export type ContainerFullWidthParams = {
-  params?: {
-    excludeTopMargin?: string;
-  };
-};
