--- examples/kit-nextjs-article-starter/src/components/container/container.props.tsx
@@ -1,5 +1,0 @@
-export type ContainerPlaceHolderProps = {
-  dynamicKey: string;
-  genericKey: string;
-  fragment: string;
-};
